+++ database/migrations/2019_11_22_175424_create_column_seller_id_in_invoices.php
@@ -9,7 +9,7 @@ class CreateColumnSellerIdInInvoices extends Migration
     public function up()
     {
         Schema::table('invoices', function (Blueprint $table) {
-            $table->unsignedBigInteger('seller_id')->after('receipt_date');
+            $table->unsignedInteger('seller_id')->after('receipt_date');
             $table->foreign('seller_id')
                 ->references('id')->on('sellers')
                 ->onDelete('cascade')
